Roboter should now stop if the distance is lower than 20cm

+++ Code/miniRobotRC/miniRobotRC.ino
@@ -42,7 +42,7 @@ uint16_t driveTimeout = 0;
 int16_t distance;
 int16_t temperature = 0;
 
-int16_t temp_time = millis();
+unsigned long temp_time = millis();
 
 //Funk
 #include <SPI.h>
